[css-grid] Apply content alignment during second pass
When we integrated content alignment in the track sizing algorithm
(r566412) we forgot to do it too when the 2nd pass is needed.

The patch is very simple and just adds the calls to apply
content alignment in LayoutGrid::RepeatTracksSizingIfNeeded()
for each axis.

The spec also mentions it in the new 3 and 4 steps:
https://drafts.csswg.org/css-grid/#algo-overview
So the new code just aligns with the spec.

BUG=870634
TEST=external/wpt/css/css-grid/alignment/grid-content-alignment-second-pass-001.html

Change-Id: Iccbcea5da1011eae37b4923bb78334cddc76d027