[LayoutNG] Pass constraint space to inline children for min/max.
Orthogonal flow root children are going to need to calculate an
extrinsic constraint space. We already got this right for direct block
children, but not for floated children of inline child nodes.

Three tests start to pass, but I wrote a new test as well, that's more
to the point (it used to fail a DCHECK, and then render incorrectly, if
we bypassed that).

Cq-Include-Trybots: luci.chromium.try:linux_layout_tests_layout_ng
Change-Id: Iffea4e8cb26f41f8b5caac4bbb7185df5f62f7d7